requirements.txt
# Web 框架
fastapi==0.122.0
uvicorn==0.38.0
pydantic==2.12.5
pydantic-settings==2.12.0

# HTTP 客户端
httpx==0.28.1
openai==2.8.1

# 机器学习
scikit-learn==1.7.2
joblib==1.5.2
fasttext-wheel==0.9.2
numpy<2.0  # fastText 尚未完全兼容 NumPy 2.0

# 中文分词
jieba==0.42.1

# 数据库
aiosqlite==0.21.0

# 系统监控
psutil==5.9.8